rewirte lipnet training
Former-commit-id: 8678df441dcd4aeb566c325b2d2ddc2d7f971b86

+++ core/utils/label.py
@@ -12,7 +12,7 @@ def text_to_labels(text: str) -> Labels:
         elif char == " ":
             ret.append(26)
 
-    return np.array(ret, dtype=np.int16)
+    return np.array(ret, dtype=np.uint8)
 
 
 def labels_to_text(labels: Labels) -> str:
